update service to set status

+++ src/main/java/codes/harsha/co2stat/adapter/in/MesurementPayload.java
@@ -7,7 +7,7 @@ import javax.validation.constraints.Pattern;
 public class MesurementPayload {
 
     @NotNull
-    public Double co2;
+    public Integer co2;
 
     @NotNull
     @NotEmpty
