+++ src/main/java/codes/harsha/co2stat/application/port/in/CollectMesurement.java
@@ -3,5 +3,5 @@ package codes.harsha.co2stat.application.port.in;
 import java.time.ZonedDateTime;
 
 public interface CollectMesurement {
-    void collect(String sensorId, double co2Level, String dateTime);
+    void collect(String sensorId, int co2Level, String dateTime);
 }
